pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iot-micro-leasing"
version = "1.0.0"
requires-python = ">=3.12"
description = "IoT Micro-Leasing Platform"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.10.2",
    "pydantic-settings==2.6.1",
    "sqlmodel==0.0.22",
    "fastapi-users[sqlmodel]==13.0.0",
    "slowapi==0.1.9",
    "redis==5.2.1",
    "structlog==24.4.0",
    "asyncpg==0.30.0",
    "django==5.1.3",
    "django-unfold==0.40.0",
    "psycopg2-binary==2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "admin_panel*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"